chore: get rid of RPC calls
